% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_deployment}
\alias{create_deployment}
\title{Creates a Deployment resource, which makes a specified RestApi callable over the internet}
\usage{
create_deployment(restApiId, stageName = NULL, stageDescription = NULL,
  description = NULL, cacheClusterEnabled = NULL, cacheClusterSize = NULL,
  variables = NULL, canarySettings = NULL, tracingEnabled = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{stageName}{The name of the Stage resource for the Deployment resource to create.}

\item{stageDescription}{The description of the Stage resource for the Deployment resource to create.}

\item{description}{The description for the Deployment resource to create.}

\item{cacheClusterEnabled}{Enables a cache cluster for the Stage resource specified in the input.}

\item{cacheClusterSize}{Specifies the cache cluster size for the Stage resource specified in the input, if a cache cluster is enabled.}

\item{variables}{A map that defines the stage variables for the Stage resource that is associated with the new deployment. Variable names can have alphanumeric and underscore characters, and the values must match \code{[A-Za-z0-9-._~:/?#&=,]+}.}

\item{canarySettings}{The input configuration for the canary deployment when the deployment is a canary release deployment.}

\item{tracingEnabled}{Specifies whether active tracing with X-ray is enabled for the Stage.}
}
\description{
Creates a Deployment resource, which makes a specified RestApi callable over the internet.
}
\examples{

}
