% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_model}
\alias{create_model}
\title{Adds a new Model resource to an existing RestApi resource}
\usage{
create_model(restApiId, name, contentType, description = NULL,
  schema = NULL)
}
\arguments{
\item{restApiId}{[Required] The RestApi identifier under which the Model will be created.}

\item{name}{[Required] The name of the model. Must be alphanumeric.}

\item{contentType}{[Required] The content-type for the model.}

\item{description}{The description of the model.}

\item{schema}{The schema for the model. For \code{application/json} models, this should be \href{%5B%5BJsonSchemaUrl%5D%5D}{JSON schema draft 4} model.}
}
\description{
Adds a new Model resource to an existing RestApi resource.
}
\examples{

}
