% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_resource}
\alias{create_resource}
\title{Creates a Resource resource}
\usage{
create_resource(restApiId, parentId, pathPart)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{parentId}{[Required] The parent resource's identifier.}

\item{pathPart}{The last path segment for this resource.}
}
\description{
Creates a Resource resource.
}
\examples{

}
