% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_usage_plan}
\alias{create_usage_plan}
\title{Creates a usage plan with the throttle and quota limits, as well as the associated API stages, specified in the payload}
\usage{
create_usage_plan(name, description = NULL, apiStages = NULL,
  throttle = NULL, quota = NULL)
}
\arguments{
\item{name}{[Required] The name of the usage plan.}

\item{description}{The description of the usage plan.}

\item{apiStages}{The associated API stages of the usage plan.}

\item{throttle}{The throttling limits of the usage plan.}

\item{quota}{The quota of the usage plan.}
}
\description{
Creates a usage plan with the throttle and quota limits, as well as the associated API stages, specified in the payload.
}
\examples{

}
