% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_usage_plan_key}
\alias{create_usage_plan_key}
\title{Creates a usage plan key for adding an existing API key to a usage plan}
\usage{
create_usage_plan_key(usagePlanId, keyId, keyType)
}
\arguments{
\item{usagePlanId}{[required] [Required] The Id of the UsagePlan resource representing the usage plan containing the to-be-created UsagePlanKey resource representing a plan customer.}

\item{keyId}{[required] [Required] The identifier of a UsagePlanKey resource for a plan customer.}

\item{keyType}{[required] [Required] The type of a UsagePlanKey resource for a plan customer.}
}
\description{
Creates a usage plan key for adding an existing API key to a usage plan.
}
\section{Accepted Parameters}{
\preformatted{create_usage_plan_key(
  usagePlanId = "string",
  keyId = "string",
  keyType = "string"
)
}
}

