% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_vpc_link}
\alias{create_vpc_link}
\title{Creates a VPC link, under the caller's account in a selected region, in an asynchronous operation that typically takes 2-4 minutes to complete and become operational}
\usage{
create_vpc_link(name, targetArns, description = NULL)
}
\arguments{
\item{name}{[Required] The name used to label and identify the VPC link.}

\item{targetArns}{[Required] The ARNs of network load balancers of the VPC targeted by the VPC link. The network load balancers must be owned by the same AWS account of the API owner.}

\item{description}{The description of the VPC link.}
}
\description{
Creates a VPC link, under the caller's account in a selected region, in an asynchronous operation that typically takes 2-4 minutes to complete and become operational. The caller must have permissions to create and update VPC Endpoint services.
}
\examples{

}
