% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_api_key}
\alias{delete_api_key}
\title{Deletes the ApiKey resource}
\usage{
delete_api_key(apiKey)
}
\arguments{
\item{apiKey}{[required] [Required] The identifier of the ApiKey resource to be deleted.}
}
\description{
Deletes the ApiKey resource.
}
\section{Accepted Parameters}{
\preformatted{delete_api_key(
  apiKey = "string"
)
}
}

