% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_authorizer}
\alias{delete_authorizer}
\title{Deletes an existing Authorizer resource}
\usage{
delete_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[required] [Required] The identifier of the Authorizer resource.}
}
\description{
Deletes an existing Authorizer resource.
}
\details{
::: {.seeAlso}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/delete-authorizer.html}{AWS CLI}
:::
}
\section{Accepted Parameters}{
\preformatted{delete_authorizer(
  restApiId = "string",
  authorizerId = "string"
)
}
}

