% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_client_certificate}
\alias{delete_client_certificate}
\title{Deletes the ClientCertificate resource}
\usage{
delete_client_certificate(
  clientCertificateId = "string"
)
}
\arguments{
\item{clientCertificateId}{[required] [Required] The identifier of the ClientCertificate resource to be deleted.}
}
\description{
Deletes the ClientCertificate resource.
}
\examples{

}
