% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_domain_name}
\alias{delete_domain_name}
\title{Deletes the DomainName resource}
\usage{
delete_domain_name(
  domainName = "string"
)
}
\arguments{
\item{domainName}{[required] [Required] The name of the DomainName resource to be deleted.}
}
\description{
Deletes the DomainName resource.
}
\examples{

}
