% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_gateway_response}
\alias{delete_gateway_response}
\title{Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings}
\usage{
delete_gateway_response(restApiId, responseType)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{responseType}{[Required]

The response type of the associated GatewayResponse. Valid values are
\itemize{
\item ACCESS\_DENIED
\item API\_CONFIGURATION\_ERROR
\item AUTHORIZER\_FAILURE
\item AUTHORIZER\_CONFIGURATION\_ERROR
\item BAD\_REQUEST\_PARAMETERS
\item BAD\_REQUEST\_BODY
\item DEFAULT\_4XX
\item DEFAULT\_5XX
\item EXPIRED\_TOKEN
\item INVALID\_SIGNATURE
\item INTEGRATION\_FAILURE
\item INTEGRATION\_TIMEOUT
\item INVALID\_API\_KEY
\item MISSING\_AUTHENTICATION\_TOKEN
\item QUOTA\_EXCEEDED
\item REQUEST\_TOO\_LARGE
\item RESOURCE\_NOT\_FOUND
\item THROTTLED
\item UNAUTHORIZED
\item UNSUPPORTED\_MEDIA\_TYPE
}}
}
\description{
Clears any customization of a GatewayResponse of a specified response type on the given RestApi and resets it with the default settings.
}
\examples{

}
