% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_integration}
\alias{delete_integration}
\title{Represents a delete integration}
\usage{
delete_integration(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] Specifies a delete integration request's resource identifier.}

\item{httpMethod}{[Required] Specifies a delete integration request's HTTP method.}
}
\description{
Represents a delete integration.
}
\examples{

}
