% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_integration_response}
\alias{delete_integration_response}
\title{Represents a delete integration response}
\usage{
delete_integration_response(restApiId, resourceId, httpMethod, statusCode)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] Specifies a delete integration response request's resource identifier.}

\item{httpMethod}{[Required] Specifies a delete integration response request's HTTP method.}

\item{statusCode}{[Required] Specifies a delete integration response request's status code.}
}
\description{
Represents a delete integration response.
}
\examples{

}
