% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_method}
\alias{delete_method}
\title{Deletes an existing Method resource}
\usage{
delete_method(restApiId, resourceId, httpMethod)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The Resource identifier for the Method resource.}

\item{httpMethod}{[Required] The HTTP verb of the Method resource.}
}
\description{
Deletes an existing Method resource.
}
\examples{

}
