% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_request_validator}
\alias{delete_request_validator}
\title{Deletes a RequestValidator of a given RestApi}
\usage{
delete_request_validator(restApiId, requestValidatorId)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[Required] The identifier of the RequestValidator to be deleted.}
}
\description{
Deletes a RequestValidator of a given RestApi.
}
\examples{

}
