% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_rest_api}
\alias{delete_rest_api}
\title{Deletes the specified API}
\usage{
delete_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}
}
\description{
Deletes the specified API.
}
\section{Accepted Parameters}{
\preformatted{delete_rest_api(
  restApiId = "string"
)
}
}

