% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_stage}
\alias{delete_stage}
\title{Deletes a Stage resource}
\usage{
delete_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{stageName}{[Required] The name of the Stage resource to delete.}
}
\description{
Deletes a Stage resource.
}
\examples{

}
