% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_usage_plan}
\alias{delete_usage_plan}
\title{Deletes a usage plan of a given plan Id}
\usage{
delete_usage_plan(usagePlanId)
}
\arguments{
\item{usagePlanId}{[Required] The Id of the to-be-deleted usage plan.}
}
\description{
Deletes a usage plan of a given plan Id.
}
\examples{

}
