% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_usage_plan_key}
\alias{delete_usage_plan_key}
\title{Deletes a usage plan key and remove the underlying API key from the associated usage plan}
\usage{
delete_usage_plan_key(usagePlanId, keyId)
}
\arguments{
\item{usagePlanId}{[Required] The Id of the UsagePlan resource representing the usage plan containing the to-be-deleted UsagePlanKey resource representing a plan customer.}

\item{keyId}{[Required] The Id of the UsagePlanKey resource to be deleted.}
}
\description{
Deletes a usage plan key and remove the underlying API key from the associated usage plan.
}
\examples{

}
