% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_vpc_link}
\alias{delete_vpc_link}
\title{Deletes an existing VpcLink of a specified identifier}
\usage{
delete_vpc_link(vpcLinkId)
}
\arguments{
\item{vpcLinkId}{[required] [Required] The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.}
}
\description{
Deletes an existing VpcLink of a specified identifier.
}
\section{Accepted Parameters}{
\preformatted{delete_vpc_link(
  vpcLinkId = "string"
)
}
}

