% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{flush_stage_authorizers_cache}
\alias{flush_stage_authorizers_cache}
\title{Flushes all authorizer cache entries on a stage}
\usage{
flush_stage_authorizers_cache(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] The string identifier of the associated RestApi.}

\item{stageName}{[required] The name of the stage to flush.}
}
\description{
Flushes all authorizer cache entries on a stage.
}
\section{Accepted Parameters}{
\preformatted{flush_stage_authorizers_cache(
  restApiId = "string",
  stageName = "string"
)
}
}

