% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{flush_stage_cache}
\alias{flush_stage_cache}
\title{Flushes a stage's cache}
\usage{
flush_stage_cache(restApiId, stageName)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{stageName}{[Required] The name of the stage to flush its cache.}
}
\description{
Flushes a stage's cache.
}
\examples{

}
