% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{generate_client_certificate}
\alias{generate_client_certificate}
\title{Generates a ClientCertificate resource}
\usage{
generate_client_certificate(description = NULL)
}
\arguments{
\item{description}{The description of the ClientCertificate.}
}
\description{
Generates a ClientCertificate resource.
}
\section{Accepted Parameters}{
\preformatted{generate_client_certificate(
  description = "string"
)
}
}

