% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_account}
\alias{get_account}
\title{Gets information about the current Account resource}
\usage{
get_account()
}
\description{
Gets information about the current Account resource.
}
\examples{

}
