% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_api_key}
\alias{get_api_key}
\title{Gets information about the current ApiKey resource}
\usage{
get_api_key(apiKey, includeValue = NULL)
}
\arguments{
\item{apiKey}{[Required] The identifier of the ApiKey resource.}

\item{includeValue}{A boolean flag to specify whether (\code{true}) or not (\code{false}) the result contains the key value.}
}
\description{
Gets information about the current ApiKey resource.
}
\examples{

}
