% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_authorizer}
\alias{get_authorizer}
\title{Describe an existing Authorizer resource}
\usage{
get_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[Required] The identifier of the Authorizer resource.}
}
\description{
Describe an existing Authorizer resource.
}
\examples{

}
