% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_authorizers}
\alias{get_authorizers}
\title{Describe an existing Authorizers resource}
\usage{
get_authorizers(restApiId, position = NULL, limit = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Describe an existing Authorizers resource.
}
\examples{

}
