% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_base_path_mappings}
\alias{get_base_path_mappings}
\title{Represents a collection of BasePathMapping resources}
\usage{
get_base_path_mappings(domainName, position = NULL, limit = NULL)
}
\arguments{
\item{domainName}{[Required] The domain name of a BasePathMapping resource.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Represents a collection of BasePathMapping resources.
}
\examples{

}
