% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_client_certificate}
\alias{get_client_certificate}
\title{Gets information about the current ClientCertificate resource}
\usage{
get_client_certificate(
  clientCertificateId = "string"
)
}
\arguments{
\item{clientCertificateId}{[required] [Required] The identifier of the ClientCertificate resource to be described.}
}
\description{
Gets information about the current ClientCertificate resource.
}
\examples{

}
