% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_client_certificates}
\alias{get_client_certificates}
\title{Gets a collection of ClientCertificate resources}
\usage{
get_client_certificates(position = NULL, limit = NULL)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Gets a collection of ClientCertificate resources.
}
\section{Accepted Parameters}{
\preformatted{get_client_certificates(
  position = "string",
  limit = 123
)
}
}

