% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_domain_name}
\alias{get_domain_name}
\title{Represents a domain name that is contained in a simpler, more intuitive URL that can be called}
\usage{
get_domain_name(domainName)
}
\arguments{
\item{domainName}{[Required] The name of the DomainName resource.}
}
\description{
Represents a domain name that is contained in a simpler, more intuitive URL that can be called.
}
\examples{

}
