% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_domain_names}
\alias{get_domain_names}
\title{Represents a collection of DomainName resources}
\usage{
get_domain_names(
  position = "string",
  limit = 123
)
}
\arguments{
\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Represents a collection of DomainName resources.
}
\examples{

}
