% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_integration}
\alias{get_integration}
\title{Get the integration settings}
\usage{
get_integration(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] Specifies a get integration request's resource identifier}

\item{httpMethod}{[required] [Required] Specifies a get integration request's HTTP method.}
}
\description{
Get the integration settings.
}
\examples{

}
