% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_integration_response}
\alias{get_integration_response}
\title{Represents a get integration response}
\usage{
get_integration_response(restApiId, resourceId, httpMethod, statusCode)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] Specifies a get integration response request's resource identifier.}

\item{httpMethod}{[Required] Specifies a get integration response request's HTTP method.}

\item{statusCode}{[Required] Specifies a get integration response request's status code.}
}
\description{
Represents a get integration response.
}
\examples{

}
