% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_method}
\alias{get_method}
\title{Describe an existing Method resource}
\usage{
get_method(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string"
)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] The Resource identifier for the Method resource.}

\item{httpMethod}{[required] [Required] Specifies the method request's HTTP method type.}
}
\description{
Describe an existing Method resource.
}
\examples{

}
