% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_method_response}
\alias{get_method_response}
\title{Describes a MethodResponse resource}
\usage{
get_method_response(restApiId, resourceId, httpMethod, statusCode)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[Required] The HTTP verb of the Method resource.}

\item{statusCode}{[Required] The status code for the MethodResponse resource.}
}
\description{
Describes a MethodResponse resource.
}
\examples{

}
