% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_model}
\alias{get_model}
\title{Describes an existing model defined for a RestApi resource}
\usage{
get_model(restApiId, modelName, flatten = NULL)
}
\arguments{
\item{restApiId}{[Required] The RestApi identifier under which the Model exists.}

\item{modelName}{[Required] The name of the model as an identifier.}

\item{flatten}{A query parameter of a Boolean value to resolve (\code{true}) all external model references and returns a flattened model schema or not (\code{false}) The default is \code{false}.}
}
\description{
Describes an existing model defined for a RestApi resource.
}
\examples{

}
