% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_request_validator}
\alias{get_request_validator}
\title{Gets a RequestValidator of a given RestApi}
\usage{
get_request_validator(restApiId, requestValidatorId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[required] [Required] The identifier of the RequestValidator to be retrieved.}
}
\description{
Gets a RequestValidator of a given RestApi.
}
\section{Accepted Parameters}{
\preformatted{get_request_validator(
  restApiId = "string",
  requestValidatorId = "string"
)
}
}

