% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_rest_api}
\alias{get_rest_api}
\title{Lists the RestApi resource in the collection}
\usage{
get_rest_api(restApiId)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}
}
\description{
Lists the RestApi resource in the collection.
}
\section{Accepted Parameters}{
\preformatted{get_rest_api(
  restApiId = "string"
)
}
}

