% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_stages}
\alias{get_stages}
\title{Gets information about one or more Stage resources}
\usage{
get_stages(restApiId, deploymentId = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{deploymentId}{The stages' deployment identifiers.}
}
\description{
Gets information about one or more Stage resources.
}
\examples{

}
