% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_tags}
\alias{get_tags}
\title{Gets the Tags collection for a given resource}
\usage{
get_tags(resourceArn, position = NULL, limit = NULL)
}
\arguments{
\item{resourceArn}{[Required] The ARN of a resource that can be tagged. The resource ARN must be URL-encoded. At present, Stage is the only taggable resource.}

\item{position}{(Not currently supported) The current pagination position in the paged result set.}

\item{limit}{(Not currently supported) The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Gets the Tags collection for a given resource.
}
\examples{

}
