% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_usage}
\alias{get_usage}
\title{Gets the usage data of a usage plan in a specified time interval}
\usage{
get_usage(usagePlanId, startDate, endDate, keyId = NULL, position = NULL,
  limit = NULL)
}
\arguments{
\item{usagePlanId}{[Required] The Id of the usage plan associated with the usage data.}

\item{startDate}{[Required] The starting date (e.g., 2016-01-01) of the usage data.}

\item{endDate}{[Required] The ending date (e.g., 2016-12-31) of the usage data.}

\item{keyId}{The Id of the API key associated with the resultant usage data.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Gets the usage data of a usage plan in a specified time interval.
}
\examples{

}
