% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_usage_plan}
\alias{get_usage_plan}
\title{Gets a usage plan of a given plan identifier}
\usage{
get_usage_plan(usagePlanId)
}
\arguments{
\item{usagePlanId}{[required] [Required] The identifier of the UsagePlan resource to be retrieved.}
}
\description{
Gets a usage plan of a given plan identifier.
}
\section{Accepted Parameters}{
\preformatted{get_usage_plan(
  usagePlanId = "string"
)
}
}

