% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_usage_plan_key}
\alias{get_usage_plan_key}
\title{Gets a usage plan key of a given key identifier}
\usage{
get_usage_plan_key(usagePlanId, keyId)
}
\arguments{
\item{usagePlanId}{[Required] The Id of the UsagePlan resource representing the usage plan containing the to-be-retrieved UsagePlanKey resource representing a plan customer.}

\item{keyId}{[Required] The key Id of the to-be-retrieved UsagePlanKey resource representing a plan customer.}
}
\description{
Gets a usage plan key of a given key identifier.
}
\examples{

}
