% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_vpc_link}
\alias{get_vpc_link}
\title{Gets a specified VPC link under the caller's account in a region}
\usage{
get_vpc_link(vpcLinkId)
}
\arguments{
\item{vpcLinkId}{[Required] The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.}
}
\description{
Gets a specified VPC link under the caller's account in a region.
}
\examples{

}
