% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{import_rest_api}
\alias{import_rest_api}
\title{A feature of the API Gateway control service for creating a new API from an external API definition file}
\usage{
import_rest_api(body, failOnWarnings = NULL, parameters = NULL)
}
\arguments{
\item{body}{[Required] The POST request body containing external API definitions. Currently, only Swagger definition JSON files are supported. The maximum size of the API definition file is 2MB.}

\item{failOnWarnings}{A query parameter to indicate whether to rollback the API creation (\code{true}) or not (\code{false}) when a warning is encountered. The default value is \code{false}.}

\item{parameters}{A key-value map of context-specific query string parameters specifying the behavior of different API importing operations. The following shows operation-specific parameters and their supported values.

To exclude DocumentationParts from the import, set \code{parameters} as \code{ignore=documentation}.

To configure the endpoint type, set \code{parameters} as \code{endpointConfigurationTypes=EDGE}, \code{endpointConfigurationTypes=REGIONAL}, or \code{endpointConfigurationTypes=PRIVATE}. The default endpoint type is \code{EDGE}.

To handle imported \code{basePath}, set \code{parameters} as \code{basePath=ignore}, \code{basePath=prepend} or \code{basePath=split}.

For example, the AWS CLI command to exclude documentation from the imported API is:\preformatted{aws apigateway import-rest-api --parameters ignore=documentation --body 'file:///path/to/imported-api-body.json'
}

The AWS CLI command to set the regional endpoint on the imported API is:\preformatted{aws apigateway import-rest-api --parameters endpointConfigurationTypes=REGIONAL --body 'file:///path/to/imported-api-body.json'
}}
}
\description{
A feature of the API Gateway control service for creating a new API from an external API definition file.
}
\examples{

}
