% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes a tag from a given resource}
\usage{
untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
\arguments{
\item{resourceArn}{[required] [Required] The ARN of a resource that can be tagged. The resource ARN must be URL-encoded. At present, Stage is the only taggable resource.}

\item{tagKeys}{[required] [Required] The Tag keys to delete.}
}
\description{
Removes a tag from a given resource.
}
\examples{

}
