% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_authorizer}
\alias{update_authorizer}
\title{Updates an existing Authorizer resource}
\usage{
update_authorizer(restApiId, authorizerId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[Required] The identifier of the Authorizer resource.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates an existing Authorizer resource.
}
\examples{

}
