% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_base_path_mapping}
\alias{update_base_path_mapping}
\title{Changes information about the BasePathMapping resource}
\usage{
update_base_path_mapping(domainName, basePath, patchOperations = NULL)
}
\arguments{
\item{domainName}{[Required] The domain name of the BasePathMapping resource to change.}

\item{basePath}{[Required] The base path of the BasePathMapping resource to change.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about the BasePathMapping resource.
}
\examples{

}
