% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_client_certificate}
\alias{update_client_certificate}
\title{Changes information about an ClientCertificate resource}
\usage{
update_client_certificate(clientCertificateId, patchOperations = NULL)
}
\arguments{
\item{clientCertificateId}{[Required] The identifier of the ClientCertificate resource to be updated.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about an ClientCertificate resource.
}
\examples{

}
