% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_deployment}
\alias{update_deployment}
\title{Changes information about a Deployment resource}
\usage{
update_deployment(restApiId, deploymentId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{deploymentId}{The replacement identifier for the Deployment resource to change information about.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about a Deployment resource.
}
\examples{

}
