% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_domain_name}
\alias{update_domain_name}
\title{Changes information about the DomainName resource}
\usage{
update_domain_name(domainName, patchOperations = NULL)
}
\arguments{
\item{domainName}{[Required] The name of the DomainName resource to be changed.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about the DomainName resource.
}
\examples{

}
