% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_integration}
\alias{update_integration}
\title{Represents an update integration}
\usage{
update_integration(restApiId, resourceId, httpMethod, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] Represents an update integration request's resource identifier.}

\item{httpMethod}{[Required] Represents an update integration request's HTTP method.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Represents an update integration.
}
\examples{

}
