% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_integration_response}
\alias{update_integration_response}
\title{Represents an update integration response}
\usage{
update_integration_response(restApiId, resourceId, httpMethod, statusCode,
  patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] Specifies an update integration response request's resource identifier.}

\item{httpMethod}{[Required] Specifies an update integration response request's HTTP method.}

\item{statusCode}{[Required] Specifies an update integration response request's status code.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Represents an update integration response.
}
\examples{

}
