% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_model}
\alias{update_model}
\title{Changes information about a model}
\usage{
update_model(restApiId, modelName, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{modelName}{[required] [Required] The name of the model to update.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about a model.
}
\section{Accepted Parameters}{
\preformatted{update_model(
  restApiId = "string",
  modelName = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

