% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_request_validator}
\alias{update_request_validator}
\title{Updates a RequestValidator of a given RestApi}
\usage{
update_request_validator(restApiId, requestValidatorId,
  patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{requestValidatorId}{[Required] The identifier of RequestValidator to be updated.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates a RequestValidator of a given RestApi.
}
\examples{

}
