% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_resource}
\alias{update_resource}
\title{Changes information about a Resource resource}
\usage{
update_resource(restApiId, resourceId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The identifier of the Resource resource.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about a Resource resource.
}
\examples{

}
