% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_rest_api}
\alias{update_rest_api}
\title{Changes information about the specified API}
\usage{
update_rest_api(restApiId, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about the specified API.
}
\examples{

}
