% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_usage_plan}
\alias{update_usage_plan}
\title{Updates a usage plan of a given plan Id}
\usage{
update_usage_plan(usagePlanId, patchOperations = NULL)
}
\arguments{
\item{usagePlanId}{[Required] The Id of the to-be-updated usage plan.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates a usage plan of a given plan Id.
}
\examples{

}
