% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_vpc_link}
\alias{update_vpc_link}
\title{Updates an existing VpcLink of a specified identifier}
\usage{
update_vpc_link(vpcLinkId, patchOperations = NULL)
}
\arguments{
\item{vpcLinkId}{[Required] The identifier of the VpcLink. It is used in an Integration to reference this VpcLink.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates an existing VpcLink of a specified identifier.
}
\examples{

}
