% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_api_key}
\alias{create_api_key}
\title{Create an ApiKey resource}
\usage{
create_api_key(
  name = "string",
  description = "string",
  enabled = TRUE|FALSE,
  generateDistinctId = TRUE|FALSE,
  value = "string",
  stageKeys = list(
    list(
      restApiId = "string",
      stageName = "string"
    )
  ),
  customerId = "string"
)
}
\arguments{
\item{name}{The name of the ApiKey.}

\item{description}{The description of the ApiKey.}

\item{enabled}{Specifies whether the ApiKey can be used by callers.}

\item{generateDistinctId}{Specifies whether (\code{true}) or not (\code{false}) the key identifier is distinct from the created API key value.}

\item{value}{Specifies a value of the API key.}

\item{stageKeys}{DEPRECATED FOR USAGE PLANS - Specifies stages associated with the API key.}

\item{customerId}{An AWS Marketplace customer identifier , when integrating with the AWS SaaS Marketplace.}
}
\description{
Create an ApiKey resource.
}
\details{
::: {.seeAlso}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/create-api-key.html}{AWS CLI}
:::
}
\examples{

}
