% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{create_resource}
\alias{create_resource}
\title{Creates a Resource resource}
\usage{
create_resource(
  restApiId = "string",
  parentId = "string",
  pathPart = "string"
)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{parentId}{[required] [Required] The parent resource's identifier.}

\item{pathPart}{[required] The last path segment for this resource.}
}
\description{
Creates a Resource resource.
}
\examples{

}
