% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_authorizer}
\alias{delete_authorizer}
\title{Deletes an existing Authorizer resource}
\usage{
delete_authorizer(restApiId, authorizerId)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{authorizerId}{[Required] The identifier of the Authorizer resource.}
}
\description{
Deletes an existing Authorizer resource.
}
\examples{

}
