% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_base_path_mapping}
\alias{delete_base_path_mapping}
\title{Deletes the BasePathMapping resource}
\usage{
delete_base_path_mapping(
  domainName = "string",
  basePath = "string"
)
}
\arguments{
\item{domainName}{[required] [Required] The domain name of the BasePathMapping resource to delete.}

\item{basePath}{[required] [Required] The base path name of the BasePathMapping resource to delete.}
}
\description{
Deletes the BasePathMapping resource.
}
\examples{

}
