% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_deployment}
\alias{delete_deployment}
\title{Deletes a Deployment resource}
\usage{
delete_deployment(restApiId, deploymentId)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{deploymentId}{[Required] The identifier of the Deployment resource to delete.}
}
\description{
Deletes a Deployment resource. Deleting a deployment will only succeed if there are no Stage resources associated with it.
}
\examples{

}
