% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_integration_response}
\alias{delete_integration_response}
\title{Represents a delete integration response}
\usage{
delete_integration_response(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string",
  statusCode = "string"
)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] Specifies a delete integration response request's resource identifier.}

\item{httpMethod}{[required] [Required] Specifies a delete integration response request's HTTP method.}

\item{statusCode}{[required] [Required] Specifies a delete integration response request's status code.}
}
\description{
Represents a delete integration response.
}
\examples{

}
