% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_model}
\alias{delete_model}
\title{Deletes a model}
\usage{
delete_model(restApiId, modelName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{modelName}{[required] [Required] The name of the model to delete.}
}
\description{
Deletes a model.
}
\section{Accepted Parameters}{
\preformatted{delete_model(
  restApiId = "string",
  modelName = "string"
)
}
}

