% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_resource}
\alias{delete_resource}
\title{Deletes a Resource resource}
\usage{
delete_resource(restApiId, resourceId)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The identifier of the Resource resource.}
}
\description{
Deletes a Resource resource.
}
\examples{

}
