% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{delete_stage}
\alias{delete_stage}
\title{Deletes a Stage resource}
\usage{
delete_stage(restApiId, stageName)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the Stage resource to delete.}
}
\description{
Deletes a Stage resource.
}
\section{Accepted Parameters}{
\preformatted{delete_stage(
  restApiId = "string",
  stageName = "string"
)
}
}

