% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_authorizers}
\alias{get_authorizers}
\title{Describe an existing Authorizers resource}
\usage{
get_authorizers(restApiId, position = NULL, limit = NULL)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{position}{The current pagination position in the paged result set.}

\item{limit}{The maximum number of returned results per page. The default value is 25 and the maximum value is 500.}
}
\description{
Describe an existing Authorizers resource.
}
\details{
::: {.seeAlso}
\href{https://docs.aws.amazon.com/cli/latest/reference/apigateway/get-authorizers.html}{AWS CLI}
:::
}
\section{Accepted Parameters}{
\preformatted{get_authorizers(
  restApiId = "string",
  position = "string",
  limit = 123
)
}
}

