% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_integration_response}
\alias{get_integration_response}
\title{Represents a get integration response}
\usage{
get_integration_response(restApiId, resourceId, httpMethod, statusCode)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{resourceId}{[required] [Required] Specifies a get integration response request's resource identifier.}

\item{httpMethod}{[required] [Required] Specifies a get integration response request's HTTP method.}

\item{statusCode}{[required] [Required] Specifies a get integration response request's status code.}
}
\description{
Represents a get integration response.
}
\section{Accepted Parameters}{
\preformatted{get_integration_response(
  restApiId = "string",
  resourceId = "string",
  httpMethod = "string",
  statusCode = "string"
)
}
}

