% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{get_stage}
\alias{get_stage}
\title{Gets information about a Stage resource}
\usage{
get_stage(
  restApiId = "string",
  stageName = "string"
)
}
\arguments{
\item{restApiId}{[required] [Required] The string identifier of the associated RestApi.}

\item{stageName}{[required] [Required] The name of the Stage resource to get information about.}
}
\description{
Gets information about a Stage resource.
}
\examples{

}
