% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_domain_name}
\alias{update_domain_name}
\title{Changes information about the DomainName resource}
\usage{
update_domain_name(domainName, patchOperations = NULL)
}
\arguments{
\item{domainName}{[required] [Required] The name of the DomainName resource to be changed.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about the DomainName resource.
}
\section{Accepted Parameters}{
\preformatted{update_domain_name(
  domainName = "string",
  patchOperations = list(
    list(
      op = "add"|"remove"|"replace"|"move"|"copy"|"test",
      path = "string",
      value = "string",
      from = "string"
    )
  )
)
}
}

