% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_method}
\alias{update_method}
\title{Updates an existing Method resource}
\usage{
update_method(restApiId, resourceId, httpMethod, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The Resource identifier for the Method resource.}

\item{httpMethod}{[Required] The HTTP verb of the Method resource.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates an existing Method resource.
}
\examples{

}
