% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_method_response}
\alias{update_method_response}
\title{Updates an existing MethodResponse resource}
\usage{
update_method_response(restApiId, resourceId, httpMethod, statusCode,
  patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{resourceId}{[Required] The Resource identifier for the MethodResponse resource.}

\item{httpMethod}{[Required] The HTTP verb of the Method resource.}

\item{statusCode}{[Required] The status code for the MethodResponse resource.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Updates an existing MethodResponse resource.
}
\examples{

}
