% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.apigateway_operations.R
\name{update_stage}
\alias{update_stage}
\title{Changes information about a Stage resource}
\usage{
update_stage(restApiId, stageName, patchOperations = NULL)
}
\arguments{
\item{restApiId}{[Required] The string identifier of the associated RestApi.}

\item{stageName}{[Required] The name of the Stage resource to change information about.}

\item{patchOperations}{A list of update operations to be applied to the specified resource and in the order specified in this list.}
}
\description{
Changes information about a Stage resource.
}
\examples{

}
